+++ src/main/scala/sc4pac/api/logger.scala
@@ -6,6 +6,7 @@ import zio.{ZIO, Task, RIO}
 import zio.http.WebSocketFrame
 import zio.http.ChannelEvent.{Read, Unregistered, UserEvent, UserEventTriggered}
 
+import JsonData as JD
 import Resolution.DepModule
 import PromptMessage.yes
 
@@ -150,8 +151,8 @@ class WebSocketPrompter(wsChannel: zio.http.WebSocketChannel, logger: WebSocketL
     }.map(_.get)
   }
 
-  def promptForVariant(module: BareModule, label: String, values: Seq[String], descriptions: Map[String, String]): Task[String] = {
-    sendPrompt(PromptMessage.ChooseVariant(module, label, values, descriptions)).map(_.body)
+  def promptForVariant(module: BareModule, label: String, values: Seq[String], info: JD.VariantInfo): Task[String] = {
+    sendPrompt(PromptMessage.ChooseVariant(module, label, values, info)).map(_.body)
   }
 
   def confirmUpdatePlan(plan: Sc4pac.UpdatePlan): zio.Task[Boolean] = {