+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- support for descriptions of variant IDs. The new `variantInfo` field should be used instead of `variantDescriptions` in newly written metadata.
+- support for default choices of variants
+
 ### Fixed
 - added support for extracting nested .rar files (#27).
+
+### Changed
+- API upgraded to 2.3:
+  - changes to format of `/prompt/choice/update/variant` message
 
 
 ## [0.5.3] - 2025-01-26